login v5

+++ src/api/routes.py
@@ -10,6 +10,7 @@ from functools import wraps
 
 api = Blueprint('api', __name__)
 app = Flask(__name__)
+
 bcrypt = Bcrypt(app)
 
 def admin_required(fn):
@@ -26,154 +27,6 @@ def admin_required(fn):
             return jsonify({"message": "Acceso no autorizado para administradores"}), 403
     return wrapper
 
-
-
-@api.route('/especialista', methods=['GET'])
-@admin_required
-def get_especialista():
-    # Supongamos que deseas obtener todos los especialistas de la base de datos
-    especialistas = Specialist.query.all()
-    
-    # Convierte los objetos Specialist en un formato serializable
-    especialistas_serializados = [especialista.serialize() for especialista in especialistas]
-    
-    return jsonify(especialistas_serializados), 200
-
-@api.route('/especialista', methods=['POST'])
-@admin_required
-def create_especialista():
-    # Obtiene los datos del especialista desde la solicitud
-    data = request.json
-
-    if not data:
-        return jsonify({"message": "Datos no proporcionados"}), 400
-
-    # Crea un nuevo objeto Specialist
-    nuevo_especialista = Specialist(
-        nombre=data.get("nombre"),
-        apellido=data.get("apellido"),
-        email=data.get("email"),
-        profesion=data.get("profesion"),
-        area_de_desempeno=data.get("area_de_desempeno"),
-        password=data.get("password")
-    )
-    # Agrega el especialista a la base de datos
-    db.session.add(nuevo_especialista)
-    db.session.commit()
-
-    # Retorna una respuesta con el nuevo especialista creado
-    return jsonify({"message": "Especialista creado con éxito", "id": nuevo_especialista.id}), 201
-
-@api.route('/especialista/<int:id>', methods=['PUT'])
-def update_especialista(id):
-    # Obtén el especialista existente por su ID
-    especialista = Specialist.query.get(id)
-
-    # Comprueba si el especialista existe
-    if especialista is None:
-        return jsonify({"message": "Especialista no encontrado"}), 404
-
-    # Obtén los datos actualizados del especialista desde la solicitud
-    data = request.json
-
-    if not data:
-        return jsonify({"message": "Datos no proporcionados"}), 400
-
-    # Actualiza los campos del especialista con los nuevos datos
-    especialista.nombre = data.get("nombre", especialista.nombre)
-    especialista.apellido = data.get("apellido", especialista.apellido)
-    especialista.email = data.get("email", especialista.email)
-    especialista.profesion = data.get("profesion", especialista.profesion)
-    especialista.area_de_desempeno = data.get("area_de_desempeno", especialista.area_de_desempeno)
-    especialista.password = data.get("password", especialista.password)
-
-    # Guarda los cambios en la base de datos
-    db.session.commit()
-
-    return jsonify({"message": "Especialista actualizado con éxito"}), 200
-
-@api.route('/especialista/<int:id>', methods=['DELETE'])
-@admin_required
-def delete_especialista(id):
-    # Obtén el especialista por su ID
-    especialista = Specialist.query.get(id)
-
-    # Comprueba si el especialista existe
-    if especialista is None:
-        return jsonify({"message": "Especialista no encontrado"}), 404
-
-    # Elimina el especialista de la base de datos
-    db.session.delete(especialista)
-    db.session.commit()
-
-    return jsonify({"message": "Especialista eliminado con éxito"}), 200
-
-@api.route('/Project', methods=['GET'])
-@admin_required
-def get_project():
-    projects = Project.query.all()
-
-    result = list(map(lambda project:project.serialize(),projects)) 
-
-    return jsonify(result), 200
-
-@api.route('/Project', methods=['POST'])
-@admin_required
-def crete_project():
-    nameProject = request.json.get("nameProject")
-    if Project.query.filter_by(nameProject=nameProject).first() is not None:
-        return jsonify({"Error": "Proyecto ya existe"}), 400
-    theme = request.json.get("theme")
-    location = request.json.get("location")
-
-    new_project = Project(
-        nameProject=nameProject,
-        theme = theme,
-        location = location
-    )
-    db.session.add(new_project)
-    db.session.commit()
-    return jsonify({"msg": "Proyecto creado satisfactoriamente"}), 200
-
-@api.route('/Project/<int:project_id>', methods=['PUT'])
-@admin_required
-def update_project(project_id):
-    project = Project.query.get(project_id)
-
-    if not project:
-        raise APIException('Proyecto no encontrado', status_code=404)
-
-    nameProject = request.json.get("nameProject")
-    theme = request.json.get("theme")
-    location = request.json.get("location")
-
-    project.nameProject = nameProject
-    project.theme = theme
-    project.location = location
-
-    db.session.commit()
-
-    return jsonify({"msg": "Proyecto actualizado satisfactoriamente"}), 200
-
-@api.route('/Project/<int:project_id>', methods=['DELETE'])
-@admin_required
-def delete_project(project_id):
-    project = Project.query.get(project_id)
-
-    if not project:
-        raise APIException('Proyecto no encontrado', status_code=404)
-
-    db.session.delete(project)
-    db.session.commit()
-
-    return jsonify({"msg": "Proyecto eliminado satisfactoriamente"}), 200
-
-
-@api.route('/admon', methods=['GET'])
-def getadmins():
-    all_admin = Administrator.query.all()
-    results = list(map(lambda item: item.serialize(),all_admin))
-    return jsonify(results), 200
 
 @api.route('/admonreg', methods=['POST'])
 def newadmins():
@@ -217,9 +70,206 @@ def loginadmin():
     token=create_access_token(identity=adminexist.id)
     return jsonify({"msg":"su usuario y contraseña son correctos", "token":token}), 201
 
-@api.route('/admon/<int:id>', methods=['PUT'])
+@api.route('/especialista', methods=['GET'])
+@jwt_required()
 @admin_required
+
+def get_especialista():
+    id_admin = get_jwt_identity()
+    # Supongamos que deseas obtener todos los especialistas de la base de datos
+    especialistas = Specialist.query.filter_by(administrator_id=id_admin)
+    
+    # Convierte los objetos Specialist en un formato serializable
+    especialistas_serializados = [especialista.serialize() for especialista in especialistas]
+    
+    return jsonify(especialistas_serializados), 200
+
+@api.route('/especialista', methods=['POST'])
+@jwt_required()
+@admin_required
+
+def create_especialista():
+    id_admin = get_jwt_identity()
+    # Obtiene los datos del especialista desde la solicitud
+    data = request.json
+
+    if not data:
+        return jsonify({"msg": "Datos no proporcionados"}), 400
+
+    # Crea un nuevo objeto Specialist
+    nuevo_especialista = Specialist(
+        nombre=data.get("nombre"),
+        apellido=data.get("apellido"),
+        email=data.get("email"),
+        profesion=data.get("profesion"),
+        area_de_desempeno=data.get("area_de_desempeno"),
+        password=data.get("password"),
+        administrator_id= id_admin
+    )
+    # Agrega el especialista a la base de datos
+    db.session.add(nuevo_especialista)
+    db.session.commit()
+
+    # Retorna una respuesta con el nuevo especialista creado
+    return jsonify({"msg": "Especialista creado con éxito", "id": nuevo_especialista.id}), 201
+
+@api.route('/especialista/<int:id>', methods=['PUT'])
+def update_especialista(id):
+    id_admin = get_jwt_identity()
+    # Obtén el especialista existente por su ID
+    especialista = Specialist.query.get(id)
+
+    # Comprueba si el especialista existe
+    if especialista is None:
+        return jsonify({"msg": "Especialista no encontrado"}), 404
+
+    if especialista.administrator_id != id_admin:
+        return jsonify({"msg": "No tienes permiso para modificar datos de este especialista"}), 403
+    # Obtén los datos actualizados del especialista desde la solicitud
+    data = request.json
+
+    if not data:
+        return jsonify({"msg": "Datos no proporcionados"}), 400
+
+    # Actualiza los campos del especialista con los nuevos datos
+    especialista.nombre = data.get("nombre", especialista.nombre)
+    especialista.apellido = data.get("apellido", especialista.apellido)
+    especialista.email = data.get("email", especialista.email)
+    especialista.profesion = data.get("profesion", especialista.profesion)
+    especialista.area_de_desempeno = data.get("area_de_desempeno", especialista.area_de_desempeno)
+    especialista.password = data.get("password", especialista.password)
+
+    # Guarda los cambios en la base de datos
+    db.session.commit()
+
+    return jsonify({"msg": "Especialista actualizado con éxito"}), 200
+
+@api.route('/especialista/<int:id>', methods=['DELETE'])
+@jwt_required()
+@admin_required
+
+def delete_especialista(id):
+    id_admin = get_jwt_identity()
+    # Obtén el especialista por su ID
+    especialista = Specialist.query.get(id)
+
+    # Comprueba si el especialista existe
+    if especialista is None:
+        return jsonify({"msg": "Especialista no encontrado"}), 404
+
+    if especialista.administrator_id != id_admin:
+        return jsonify({"msg": "No tienes permiso para eliminar este especialista"}), 403
+    # Elimina el especialista de la base de datos
+    db.session.delete(especialista)
+    db.session.commit()
+
+    return jsonify({"msg": "Especialista eliminado con éxito"}), 200
+
+@api.route('/Project', methods=['GET'])
+@jwt_required()
+@admin_required
+
+def get_project():
+    id_admin = get_jwt_identity()
+    projects = Project.query.filter_by(admon_id=id_admin).all()
+    result = list(map(lambda project:project.serialize(),projects)) 
+    return jsonify(result), 200
+
+@api.route('/Project', methods=['POST'])
+@jwt_required()
+@admin_required
+
+def crete_project():
+    id_admin = get_jwt_identity()
+    nameProject = request.json.get("nameProject")
+    if Project.query.filter_by(nameProject=nameProject).first() is not None:
+        return jsonify({"Error": "Proyecto ya existe"}), 400
+    theme = request.json.get("theme")
+    location = request.json.get("location")
+
+    new_project = Project(
+        nameProject=nameProject,
+        theme = theme,
+        location = location,
+        admon_id=id_admin
+    )
+    db.session.add(new_project)
+    db.session.commit()
+    return jsonify({"msg": "Proyecto creado satisfactoriamente"}), 200
+
+@api.route('/Project/<int:project_id>', methods=['PUT'])
+@jwt_required()
+@admin_required
+
+def update_project(project_id):
+    id_admin = get_jwt_identity()
+    
+    project = Project.query.get(project_id)
+
+    if not project:
+        raise APIException('Proyecto no encontrado', status_code=404)
+    
+    if project.admon_id != id_admin:
+        raise APIException('No tienes permiso para actualizar este proyecto', status_code=403)
+
+    nameProject = request.json.get("nameProject")
+    theme = request.json.get("theme")
+    location = request.json.get("location")
+
+    project.nameProject = nameProject
+    project.theme = theme
+    project.location = location
+
+    admin_id = request.json.get("admin_id")
+    if admin_id is not None:
+        project.admin_id = admin_id
+
+    db.session.commit()
+    return jsonify({"msg": "Proyecto actualizado satisfactoriamente"}), 200
+
+@api.route('/Project/<int:project_id>', methods=['DELETE'])
+@jwt_required()
+@admin_required
+
+def delete_project(project_id):
+    id_admin = get_jwt_identity()
+    project = Project.query.get(project_id)
+
+    if not project:
+        raise APIException('Proyecto no encontrado', status_code=404)
+
+    if project.admon_id == id_admin:
+        raise APIException('No tienes permiso para eliminar este proyecto', status_code=403)
+    
+    db.session.delete(project)
+    db.session.commit()
+
+    return jsonify({"msg": "Proyecto eliminado satisfactoriamente"}), 200
+
+
+@api.route('/admon', methods=['GET'])
+@jwt_required()
+@admin_required
+
+def getadmins():
+    id_admin = get_jwt_identity()
+    administrator = Administrator.query.filter_by(id=id_admin).first()
+    if not administrator:
+        return jsonify({"msg": "no existe este administrador"}), 404
+    
+    results = administrator.serialize()
+    return jsonify(results), 200
+
+
+@api.route('/admon/<int:id>', methods=['PUT'])
+@jwt_required()
+@admin_required
+
 def modifadmins(id):
+    id_admin = get_jwt_identity()
+    if id != id_admin:
+        return jsonify({"msg":"No tiene permisos para modificar este administrador"}), 403 
+    
     modifadmin = Administrator.query.get(id)
     modifadmin.name = request.json.get("name")
     modifadmin.lastname = request.json.get("lastname")
@@ -229,16 +279,21 @@ def modifadmins(id):
     modifadmin.aditional_info = request.json.get("aditional_info")
     print(modifadmin)
     db.session.commit()
-    return jsonify({"msg":"Usuario modificado"}), 201 
+    return jsonify({"msg":"haz  modificado tu perfil administrador"}), 201 
 
 @api.route('/admon/<int:id>', methods=['DELETE'])
+@jwt_required()
 @admin_required
+
 def deleteadmins(id):
+    id_admin = get_jwt_identity()
+    if id != id_admin:
+        return jsonify({"msg":"No tiene permisos para modificar este administrador"}), 403 
     deleteadmin = Administrator.query.filter_by(id=id).first()
     db.session.delete(deleteadmin)
     db.session.commit()
     return jsonify({"msg":"Usuario eliminado"}), 201 
-
+
